requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

pandas==2.2.2
openpyxl==3.1.5

google-cloud-bigquery==3.25.0
pyarrow==17.0.0

pydantic-settings==2.5.2
